# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "State Machine API",
    "description": "REST API for managing state machines and executions",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/state-machines/api/v1",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/state-machines/{stateMachineId}/executions/batch": {
      "post": {
        "tags": ["Batch"],
        "summary": "Execute batch",
        "description": "Execute a batch of executions with specified parameters",
        "operationId": "executeBatch",
        "parameters": [
          {
            "$ref": "#/components/parameters/StateMachineId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch execution initiated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchExecutionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/queue/enqueue": {
      "post": {
        "tags": ["Queue"],
        "summary": "Enqueue execution",
        "description": "Add an execution task to the distributed queue",
        "operationId": "enqueueExecution",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnqueueExecutionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Execution enqueued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnqueueExecutionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "StateMachineId": {
        "name": "stateMachineId",
        "in": "path",
        "required": true,
        "description": "Unique identifier of the state machine",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["error", "code"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message"
          },
          "code": {
            "type": "integer",
            "description": "HTTP status code"
          }
        }
      },
...
      "ExecuteBatchRequest": {
        "type": "object",
        "properties": {
          "filter": {
            "$ref": "#/components/schemas/ExecutionFilterRequest"
          },
          "namePrefix": {
            "type": "string",
            "description": "Prefix for execution names in the batch"
          },
          "concurrency": {
            "type": "integer",
            "description": "Number of concurrent executions",
            "default": 10
          },
          "mode": {
            "type": "string",
            "enum": ["distributed", "concurrent", "sequential"],
            "description": "Execution mode"
          },
          "stopOnError": {
            "type": "boolean",
            "description": "Stop batch if an error occurs"
          },
          "executionNameList": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ExecutionFilterRequest": {
        "type": "object",
        "properties": {
          "stateMachineId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "startTimeFrom": {
            "type": "integer",
            "format": "int64"
          },
          "startTimeTo": {
            "type": "integer",
            "format": "int64"
          },
          "namePattern": {
            "type": "string"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "states": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "BatchExecutionResponse": {
        "type": "object",
        "required": ["batchId", "totalEnqueued", "totalFailed", "mode"],
        "properties": {
          "batchId": {
            "type": "string"
          },
          "totalEnqueued": {
            "type": "integer"
          },
          "totalFailed": {
            "type": "integer"
          },
          "mode": {
            "type": "string"
          }
        }
      },
      "EnqueueExecutionRequest": {
        "type": "object",
        "required": ["stateMachineId", "executionName"],
        "properties": {
          "stateMachineId": {
            "type": "string"
          },
          "executionName": {
            "type": "string"
          },
          "input": {
            "type": "object"
          },
          "queue": {
            "type": "string"
          },
          "sourceExecutionId": {
            "type": "string"
          },
          "sourceStateName": {
            "type": "string"
          }
        }
      },
      "EnqueueExecutionResponse": {
        "type": "object",
        "required": ["taskId", "queue", "enqueuedAt"],
        "properties": {
          "taskId": {
            "type": "string"
          },
          "queue": {
            "type": "string"
          },
          "enqueuedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}